add normal ai logic(move on diagonal, hunt move, random move)

+++ objects/ship.py
@@ -44,7 +44,7 @@ class ShipsDrawer:
                     coordinates.append([ship[deck].x_on_field, ship[deck].y_on_field])
             return coordinates
 
-    def tick_cells_around_ship(self, ship, field, double_field):
+    def tick_cells_around_ship(self, ship, field, double_field=None):
         for deck in range(len(ship)):
             for x, y in product([-1, 0, 1], repeat=2):
                 x_on_field = ship[deck].x_on_field
@@ -52,7 +52,8 @@ class ShipsDrawer:
                 if 0 <= y_on_field + y < len(field) and 0 <= x_on_field + x < len(field[0]):
                     if field[y_on_field + y][x_on_field + x].type != 3:
                         field[y_on_field + y][x_on_field + x].set_type(2)
-                        double_field[y_on_field + y][x_on_field + x].delete()
+                        if double_field is not None:
+                            double_field[y_on_field + y][x_on_field + x].delete()
 
     def find_ship(self, ships, field, x_on_field, y_on_field):
         for ship in range(len(ships)):
